Move internal/resolve to resolve
