+++ internal/language/go/BUILD.bazel
@@ -51,9 +51,9 @@ go_library(
         "//internal/language:go_default_library",
         "//internal/language/proto:go_default_library",
         "//internal/repos:go_default_library",
-        "//internal/resolve:go_default_library",
         "//label:go_default_library",
         "//pathtools:go_default_library",
+        "//resolve:go_default_library",
         "//rule:go_default_library",
         "//vendor/github.com/bazelbuild/buildtools/build:go_default_library",
     ],
@@ -77,10 +77,10 @@ go_test(
         "//internal/language/proto:go_default_library",
         "//internal/merger:go_default_library",
         "//internal/repos:go_default_library",
-        "//internal/resolve:go_default_library",
         "//internal/testtools:go_default_library",
         "//internal/walk:go_default_library",
         "//label:go_default_library",
+        "//resolve:go_default_library",
         "//rule:go_default_library",
         "//vendor/github.com/bazelbuild/buildtools/build:go_default_library",
         "//vendor/golang.org/x/tools/go/vcs:go_default_library",
