+++ internal/language/proto/BUILD.bazel
@@ -30,8 +30,8 @@ go_library(
         "//config:go_default_library",
         "//internal/language:go_default_library",
         "//internal/repos:go_default_library",
-        "//internal/resolve:go_default_library",
         "//label:go_default_library",
+        "//resolve:go_default_library",
         "//rule:go_default_library",
     ],
 )
@@ -50,10 +50,10 @@ go_test(
         "//internal/language:go_default_library",
         "//internal/merger:go_default_library",
         "//internal/repos:go_default_library",
-        "//internal/resolve:go_default_library",
         "//internal/testtools:go_default_library",
         "//internal/walk:go_default_library",
         "//label:go_default_library",
+        "//resolve:go_default_library",
         "//rule:go_default_library",
         "//vendor/github.com/bazelbuild/buildtools/build:go_default_library",
     ],
