+++ resolve/config.go
